refactor: de-part-file incremental_candidate_validator (#3543)
Promote the incremental_candidate_validator prefix-replay and economy-
projection fragments from `part of` files to a single standalone companion
library with explicit imports, continuing the orders-package de-part-file
slice of #3543.

- incremental_candidate_validator_replay.dart (new): merges the former
  *_prefix_replay.dart and *_projection.dart parts into one standalone
  library holding both probe extensions. Keeping both extensions in the same
  library lets their private helpers (_player, _ensurePostWorkPrefixState,
  etc.) stay private without widening the package surface.
- incremental_candidate_validator_cache.dart (new): the per-pass memoization
  slots formerly held as private instance fields move into a package-internal
  IncrementalCandidateValidatorCache holder, exposed via the validator's
  public `cache` field so the companion library reads/writes them through the
  public surface instead of shared part-scope.
- incremental_candidate_validator.dart re-exports both new libraries, so the
  package public surface (isBuildAccepted/isWorkAccepted/isRecruitWorkerAccepted/
  isDiplomaticAccepted/probeDiplomaticOrder) is unchanged.

Behavior-preserving (pure structural refactor): the moved probe logic is
identical and a fresh cache holder per validator instance matches the prior
per-instance caching. All files stay under the 500-line domain-source cap.
This removes 2 of the remaining 3 `part of` directives in orders lib; only
the generated order_engine.g.dart part remains.

Refs #3543
